requirements.txt
requests==2.27.1
pandas==1.3.0
streamlit==1.1.0
plotly==4.14.3
Pillow==9.0.0

